This works like I expect it to.
